move PromptIfDirty component
